requirements.txt
# Build tools
setuptools==69.2.0

# Web Framework và API
fastapi==0.104.1
uvicorn==0.24.0
starlette==0.27.0
python-multipart==0.0.6

# Database
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9

# MQTT & IoT
paho-mqtt==2.1.0
adafruit-io==2.8.0

# Auth & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
ecdsa==0.19.0
pyasn1==0.6.1
rsa==4.9
PyJWT==2.8.0

# Data Processing
numpy==1.26.2
pandas==2.1.3
scipy==1.11.4

# Visualization
matplotlib==3.8.2
seaborn==0.13.0
contourpy==1.2.0
kiwisolver==1.4.5
pillow==10.1.0

# Config & Environment
python-dotenv==1.0.0
pydantic==2.5.2
pydantic-settings==2.1.0
pydantic_core==2.14.5

# Scheduling
APScheduler==3.10.4
schedule==1.2.1
tzlocal==5.2

# Utilities
python-dateutil==2.8.2
tabulate==0.9.0
requests==2.31.0
certifi==2023.11.17

# Formatting & Compatibility
charset-normalizer==3.3.2
typing_extensions==4.8.0
annotated-types==0.6.0
anyio==3.7.1
idna==3.6
six==1.16.0
sniffio==1.3.0
h11==0.14.0
click==8.1.7
exceptiongroup==1.1.3

# Date and Time
pytz==2023.3.post1
tzdata==2023.3

# Other
fonttools==4.45.1
pyparsing==3.1.1
packaging==23.2
zipp==3.17.0
importlib_resources==6.1.1


#AI
scikit-learn